+++ src/vs/workbench/contrib/notebook/browser/controller/coreActions.ts
@@ -19,7 +19,7 @@ import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation
 import { KeybindingWeight } from 'vs/platform/keybinding/common/keybindingsRegistry';
 import { IQuickInputService, IQuickPickItem, QuickPickInput } from 'vs/platform/quickinput/common/quickInput';
 import { BaseCellRenderTemplate, CellEditState, CellFocusMode, EXECUTE_CELL_COMMAND_ID, getNotebookEditorFromEditorPane, IActiveNotebookEditor, ICellViewModel, NOTEBOOK_CELL_EDITABLE, NOTEBOOK_CELL_HAS_OUTPUTS, NOTEBOOK_CELL_LIST_FOCUSED, NOTEBOOK_CELL_MARKDOWN_EDIT_MODE, NOTEBOOK_CELL_EXECUTION_STATE, NOTEBOOK_CELL_TYPE, NOTEBOOK_EDITOR_EDITABLE, NOTEBOOK_EDITOR_FOCUSED, NOTEBOOK_IS_ACTIVE_EDITOR, NOTEBOOK_KERNEL_COUNT, NOTEBOOK_INTERRUPTIBLE_KERNEL, NOTEBOOK_HAS_RUNNING_CELL, CHANGE_CELL_LANGUAGE, QUIT_EDIT_CELL_COMMAND_ID, NOTEBOOK_USE_CONSOLIDATED_OUTPUT_BUTTON, NOTEBOOK_HAS_OUTPUTS, NOTEBOOK_CELL_EXECUTING, NOTEBOOK_MISSING_KERNEL_EXTENSION, cellRangeToViewCells } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
-import { CellEditType, CellKind, ICellEditOperation, isDocumentExcludePattern, NotebookCellExecutionState, TransientCellMetadata, TransientDocumentMetadata, SelectionStateType, ICellReplaceEdit, OpenGettingStarted, GlobalToolbarShowLabel, ConsolidatedRunButton } from 'vs/workbench/contrib/notebook/common/notebookCommon';
+import { CellEditType, CellKind, ICellEditOperation, isDocumentExcludePattern, NotebookCellExecutionState, TransientCellMetadata, TransientDocumentMetadata, SelectionStateType, ICellReplaceEdit, OpenGettingStarted, ConsolidatedRunButton } from 'vs/workbench/contrib/notebook/common/notebookCommon';
 import { ICellRange, isICellRange } from 'vs/workbench/contrib/notebook/common/notebookRange';
 import { INotebookService } from 'vs/workbench/contrib/notebook/common/notebookService';
 import { IEditorGroupsService } from 'vs/workbench/services/editor/common/editorGroupsService';
@@ -31,7 +31,7 @@ import { EditorsOrder, IEditorCommandsContext } from 'vs/workbench/common/editor
 import { INotebookEditorService } from 'vs/workbench/contrib/notebook/browser/notebookEditorService';
 import { ITelemetryService } from 'vs/platform/telemetry/common/telemetry';
 import { WorkbenchActionExecutedClassification, WorkbenchActionExecutedEvent } from 'vs/base/common/actions';
-import { CellViewModel, NotebookViewModel } from 'vs/workbench/contrib/notebook/browser/viewModel/notebookViewModel';
+import { NotebookViewModel } from 'vs/workbench/contrib/notebook/browser/viewModel/notebookViewModel';
 import { INotebookKernelService } from 'vs/workbench/contrib/notebook/common/notebookKernelService';
 import { Iterable } from 'vs/base/common/iterator';
 import { flatten, maxIndex, minIndex } from 'vs/base/common/arrays';
@@ -52,15 +52,6 @@ const CLEAR_ALL_CELLS_OUTPUTS_COMMAND_ID = 'notebook.clearAllCellsOutputs';
 const RENDER_ALL_MARKDOWN_CELLS = 'notebook.renderAllMarkdownCells';
 
 // Cell Commands
-const INSERT_CODE_CELL_ABOVE_COMMAND_ID = 'notebook.cell.insertCodeCellAbove';
-const INSERT_CODE_CELL_BELOW_COMMAND_ID = 'notebook.cell.insertCodeCellBelow';
-const INSERT_CODE_CELL_ABOVE_AND_FOCUS_CONTAINER_COMMAND_ID = 'notebook.cell.insertCodeCellAboveAndFocusContainer';
-const INSERT_CODE_CELL_BELOW_AND_FOCUS_CONTAINER_COMMAND_ID = 'notebook.cell.insertCodeCellBelowAndFocusContainer';
-const INSERT_CODE_CELL_AT_TOP_COMMAND_ID = 'notebook.cell.insertCodeCellAtTop';
-const INSERT_MARKDOWN_CELL_ABOVE_COMMAND_ID = 'notebook.cell.insertMarkdownCellAbove';
-const INSERT_MARKDOWN_CELL_BELOW_COMMAND_ID = 'notebook.cell.insertMarkdownCellBelow';
-const INSERT_MARKDOWN_CELL_AT_TOP_COMMAND_ID = 'notebook.cell.insertMarkdownCellAtTop';
-
 const EDIT_CELL_COMMAND_ID = 'notebook.cell.edit';
 const DELETE_CELL_COMMAND_ID = 'notebook.cell.delete';
 
@@ -1127,309 +1118,6 @@ export async function changeCellToKind(kind: CellKind, context: INotebookActionC
 		}, undefined, true);
 	}
 }
-
-abstract class InsertCellCommand extends NotebookAction {
-	constructor(
-		desc: Readonly<IAction2Options>,
-		private kind: CellKind,
-		private direction: 'above' | 'below',
-		private focusEditor: boolean
-	) {
-		super(desc);
-	}
-
-	async runWithContext(accessor: ServicesAccessor, context: INotebookActionContext): Promise<void> {
-		let newCell: CellViewModel | null = null;
-		if (context.ui) {
-			context.notebookEditor.focus();
-		}
-
-		if (context.cell) {
-			newCell = context.notebookEditor.insertNotebookCell(context.cell, this.kind, this.direction, undefined, true);
-		} else {
-			const focusRange = context.notebookEditor.getFocus();
-			const next = focusRange.end - 1;
-			newCell = context.notebookEditor.insertNotebookCell(context.notebookEditor.viewModel.viewCells[next], this.kind, this.direction, undefined, true);
-		}
-
-		if (newCell) {
-			context.notebookEditor.focusNotebookCell(newCell, this.focusEditor ? 'editor' : 'container');
-		}
-	}
-}
-
-registerAction2(class InsertCodeCellAboveAction extends InsertCellCommand {
-	constructor() {
-		super(
-			{
-				id: INSERT_CODE_CELL_ABOVE_COMMAND_ID,
-				title: localize('notebookActions.insertCodeCellAbove', "Insert Code Cell Above"),
-				keybinding: {
-					primary: KeyMod.CtrlCmd | KeyMod.Shift | KeyCode.Enter,
-					when: ContextKeyExpr.and(NOTEBOOK_CELL_LIST_FOCUSED, InputFocusedContext.toNegated()),
-					weight: KeybindingWeight.WorkbenchContrib
-				},
-				menu: {
-					id: MenuId.NotebookCellInsert,
-					order: 0
-				}
-			},
-			CellKind.Code,
-			'above',
-			true);
-	}
-});
-
-registerAction2(class InsertCodeCellAboveAndFocusContainerAction extends InsertCellCommand {
-	constructor() {
-		super(
-			{
-				id: INSERT_CODE_CELL_ABOVE_AND_FOCUS_CONTAINER_COMMAND_ID,
-				title: localize('notebookActions.insertCodeCellAboveAndFocusContainer', "Insert Code Cell Above and Focus Container")
-			},
-			CellKind.Code,
-			'above',
-			false);
-	}
-});
-
-registerAction2(class InsertCodeCellBelowAction extends InsertCellCommand {
-	constructor() {
-		super(
-			{
-				id: INSERT_CODE_CELL_BELOW_COMMAND_ID,
-				title: localize('notebookActions.insertCodeCellBelow', "Insert Code Cell Below"),
-				keybinding: {
-					primary: KeyMod.CtrlCmd | KeyCode.Enter,
-					when: ContextKeyExpr.and(NOTEBOOK_CELL_LIST_FOCUSED, InputFocusedContext.toNegated()),
-					weight: KeybindingWeight.WorkbenchContrib
-				},
-				menu: {
-					id: MenuId.NotebookCellInsert,
-					order: 1
-				}
-			},
-			CellKind.Code,
-			'below',
-			true);
-	}
-});
-
-registerAction2(class InsertCodeCellBelowAndFocusContainerAction extends InsertCellCommand {
-	constructor() {
-		super(
-			{
-				id: INSERT_CODE_CELL_BELOW_AND_FOCUS_CONTAINER_COMMAND_ID,
-				title: localize('notebookActions.insertCodeCellBelowAndFocusContainer', "Insert Code Cell Below and Focus Container"),
-			},
-			CellKind.Code,
-			'below',
-			false);
-	}
-});
-
-registerAction2(class InsertCodeCellAtTopAction extends NotebookAction {
-	constructor() {
-		super(
-			{
-				id: INSERT_CODE_CELL_AT_TOP_COMMAND_ID,
-				title: localize('notebookActions.insertCodeCellAtTop', "Add Code Cell At Top"),
-				f1: false
-			});
-	}
-
-	override async run(accessor: ServicesAccessor, context?: INotebookActionContext): Promise<void> {
-		context = context ?? this.getEditorContextFromArgsOrActive(accessor);
-		if (context) {
-			this.runWithContext(accessor, context);
-		}
-	}
-
-	async runWithContext(accessor: ServicesAccessor, context: INotebookActionContext): Promise<void> {
-		const newCell = context.notebookEditor.insertNotebookCell(undefined, CellKind.Code, 'above', undefined, true);
-		if (newCell) {
-			context.notebookEditor.focusNotebookCell(newCell, 'editor');
-		}
-	}
-});
-
-registerAction2(class InsertMarkdownCellAtTopAction extends NotebookAction {
-	constructor() {
-		super(
-			{
-				id: INSERT_MARKDOWN_CELL_AT_TOP_COMMAND_ID,
-				title: localize('notebookActions.insertMarkdownCellAtTop', "Add Markdown Cell At Top"),
-				f1: false
-			});
-	}
-
-	override async run(accessor: ServicesAccessor, context?: INotebookActionContext): Promise<void> {
-		context = context ?? this.getEditorContextFromArgsOrActive(accessor);
-		if (context) {
-			this.runWithContext(accessor, context);
-		}
-	}
-
-	async runWithContext(accessor: ServicesAccessor, context: INotebookActionContext): Promise<void> {
-		const newCell = context.notebookEditor.insertNotebookCell(undefined, CellKind.Markup, 'above', undefined, true);
-		if (newCell) {
-			context.notebookEditor.focusNotebookCell(newCell, 'editor');
-		}
-	}
-});
-
-MenuRegistry.appendMenuItem(MenuId.NotebookCellBetween, {
-	command: {
-		id: INSERT_CODE_CELL_BELOW_COMMAND_ID,
-		title: localize('notebookActions.menu.insertCode', "$(add) Code"),
-		tooltip: localize('notebookActions.menu.insertCode.tooltip', "Add Code Cell")
-	},
-	order: 0,
-	group: 'inline',
-	when: ContextKeyExpr.and(
-		NOTEBOOK_EDITOR_EDITABLE.isEqualTo(true),
-		ContextKeyExpr.notEquals('config.notebook.experimental.insertToolbarAlignment', 'left')
-	)
-});
-
-MenuRegistry.appendMenuItem(MenuId.NotebookCellBetween, {
-	command: {
-		id: INSERT_CODE_CELL_BELOW_COMMAND_ID,
-		title: localize('notebookActions.menu.insertCode.minimalToolbar', "Add Code"),
-		icon: Codicon.add,
-		tooltip: localize('notebookActions.menu.insertCode.tooltip', "Add Code Cell")
-	},
-	order: 0,
-	group: 'inline',
-	when: ContextKeyExpr.and(
-		NOTEBOOK_EDITOR_EDITABLE.isEqualTo(true),
-		ContextKeyExpr.equals('config.notebook.experimental.insertToolbarAlignment', 'left')
-	)
-});
-
-MenuRegistry.appendMenuItem(MenuId.NotebookToolbar, {
-	command: {
-		id: INSERT_CODE_CELL_BELOW_COMMAND_ID,
-		icon: Codicon.add,
-		title: localize('notebookActions.menu.insertCode.ontoolbar', "Code"),
-		tooltip: localize('notebookActions.menu.insertCode.tooltip', "Add Code Cell")
-	},
-	order: -5,
-	group: 'navigation/add',
-	when: ContextKeyExpr.and(
-		NOTEBOOK_EDITOR_EDITABLE.isEqualTo(true),
-		ContextKeyExpr.notEquals('config.notebook.insertToolbarLocation', 'betweenCells'),
-		ContextKeyExpr.notEquals('config.notebook.insertToolbarLocation', 'hidden')
-	)
-});
-
-MenuRegistry.appendMenuItem(MenuId.NotebookCellListTop, {
-	command: {
-		id: INSERT_CODE_CELL_AT_TOP_COMMAND_ID,
-		title: localize('notebookActions.menu.insertCode', "$(add) Code"),
-		tooltip: localize('notebookActions.menu.insertCode.tooltip', "Add Code Cell")
-	},
-	order: 0,
-	group: 'inline',
-	when: ContextKeyExpr.and(
-		NOTEBOOK_EDITOR_EDITABLE.isEqualTo(true),
-		ContextKeyExpr.notEquals('config.notebook.experimental.insertToolbarAlignment', 'left')
-	)
-});
-
-MenuRegistry.appendMenuItem(MenuId.NotebookCellListTop, {
-	command: {
-		id: INSERT_CODE_CELL_AT_TOP_COMMAND_ID,
-		title: localize('notebookActions.menu.insertCode.minimaltoolbar', "Add Code"),
-		icon: Codicon.add,
-		tooltip: localize('notebookActions.menu.insertCode.tooltip', "Add Code Cell")
-	},
-	order: 0,
-	group: 'inline',
-	when: ContextKeyExpr.and(
-		NOTEBOOK_EDITOR_EDITABLE.isEqualTo(true),
-		ContextKeyExpr.equals('config.notebook.experimental.insertToolbarAlignment', 'left')
-	)
-});
-
-registerAction2(class InsertMarkdownCellAboveAction extends InsertCellCommand {
-	constructor() {
-		super(
-			{
-				id: INSERT_MARKDOWN_CELL_ABOVE_COMMAND_ID,
-				title: localize('notebookActions.insertMarkdownCellAbove', "Insert Markdown Cell Above"),
-				menu: {
-					id: MenuId.NotebookCellInsert,
-					order: 2
-				}
-			},
-			CellKind.Markup,
-			'above',
-			true);
-	}
-});
-
-registerAction2(class InsertMarkdownCellBelowAction extends InsertCellCommand {
-	constructor() {
-		super(
-			{
-				id: INSERT_MARKDOWN_CELL_BELOW_COMMAND_ID,
-				title: localize('notebookActions.insertMarkdownCellBelow', "Insert Markdown Cell Below"),
-				menu: {
-					id: MenuId.NotebookCellInsert,
-					order: 3
-				}
-			},
-			CellKind.Markup,
-			'below',
-			true);
-	}
-});
-
-MenuRegistry.appendMenuItem(MenuId.NotebookCellBetween, {
-	command: {
-		id: INSERT_MARKDOWN_CELL_BELOW_COMMAND_ID,
-		title: localize('notebookActions.menu.insertMarkdown', "$(add) Markdown"),
-		tooltip: localize('notebookActions.menu.insertMarkdown.tooltip', "Add Markdown Cell")
-	},
-	order: 1,
-	group: 'inline',
-	when: ContextKeyExpr.and(
-		NOTEBOOK_EDITOR_EDITABLE.isEqualTo(true),
-		ContextKeyExpr.notEquals('config.notebook.experimental.insertToolbarAlignment', 'left')
-	)
-});
-
-MenuRegistry.appendMenuItem(MenuId.NotebookToolbar, {
-	command: {
-		id: INSERT_MARKDOWN_CELL_BELOW_COMMAND_ID,
-		icon: Codicon.add,
-		title: localize('notebookActions.menu.insertMarkdown.ontoolbar', "Markdown"),
-		tooltip: localize('notebookActions.menu.insertMarkdown.tooltip', "Add Markdown Cell")
-	},
-	order: -5,
-	group: 'navigation/add',
-	when: ContextKeyExpr.and(
-		NOTEBOOK_EDITOR_EDITABLE.isEqualTo(true),
-		ContextKeyExpr.notEquals('config.notebook.insertToolbarLocation', 'betweenCells'),
-		ContextKeyExpr.notEquals('config.notebook.insertToolbarLocation', 'hidden'),
-		ContextKeyExpr.notEquals(`config.${GlobalToolbarShowLabel}`, false)
-	)
-});
-
-MenuRegistry.appendMenuItem(MenuId.NotebookCellListTop, {
-	command: {
-		id: INSERT_MARKDOWN_CELL_AT_TOP_COMMAND_ID,
-		title: localize('notebookActions.menu.insertMarkdown', "$(add) Markdown"),
-		tooltip: localize('notebookActions.menu.insertMarkdown.tooltip', "Add Markdown Cell")
-	},
-	order: 1,
-	group: 'inline',
-	when: ContextKeyExpr.and(
-		NOTEBOOK_EDITOR_EDITABLE.isEqualTo(true),
-		ContextKeyExpr.notEquals('config.notebook.experimental.insertToolbarAlignment', 'left')
-	)
-});
 
 registerAction2(class EditCellAction extends NotebookCellAction {
 	constructor() {
